chore: Change directory name from conf to config (#18)
