+++ srcs/config/Config.hpp
@@ -4,18 +4,20 @@
 # define BUFFER_SIZE 1600
 
 # include "Block.hpp"
+# include <unistd.h>
+# include <fcntl.h>
 # include <vector>
 
 class Config {
 private:
 	std::vector<Block>			serverBlocks_;
 
-	std::vector<std::string> 	splitLines(const std::string &str, std::string delim);
+	Config();
+	std::vector<std::string>	splitLines(const std::string &str, std::string delim);
 	std::vector<std::string>	readAndSplit(const char *filePath);
 	void						parseConfigFile(const char *filePath);
 
 public:
-	Config();
 	Config(char *filePath);
 	~Config();
